+++ src/app.module.ts
@@ -1,13 +1,14 @@
 import { MiddlewareConsumer, Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { GracefulShutdownModule } from '@tygra/nestjs-graceful-shutdown';
+// import { GracefulShutdownModule } from '@tygra/nestjs-graceful-shutdown';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { RequestIdMiddleware } from './common/middlewares';
-import { getGracefulShutdownConfig, getTypeOrmModuleOptions } from './config';
+import { QueryLoggerMiddleware, RequestIdMiddleware } from './common/middlewares';
+import { /*getGracefulShutdownConfig,*/ getTypeOrmModuleOptions } from './config';
 import { getEnvFile, validate } from './core/environment';
+import { GraphqlModule } from './graphql/graphql.module';
 import { OrdersModule } from './orders/orders.module';
 import { ProductsModule } from './products/products.module';
 import { UsersModule } from './users/users.module';
@@ -25,15 +26,18 @@ import { UsersModule } from './users/users.module';
       inject: [ConfigService],
       useFactory: getTypeOrmModuleOptions,
     }),
-    GracefulShutdownModule.forRoot(getGracefulShutdownConfig()),
+    // TODO: Uncomment when resolve problem with graphql module
+    // GracefulShutdownModule.forRoot(getGracefulShutdownConfig()),
     UsersModule,
     OrdersModule,
     ProductsModule,
+    GraphqlModule,
   ],
   providers: [AppService],
 })
 export class AppModule {
   configure(consumer: MiddlewareConsumer) {
     consumer.apply(RequestIdMiddleware).forRoutes('*');
+    consumer.apply(QueryLoggerMiddleware).forRoutes('*');
   }
 }